+++ playwright/tests/dc_homepage.spec.ts
@@ -2,7 +2,7 @@ import { test, expect } from "@playwright/test";
 import { DCHomepage } from "../pages/dc_homepage";
 
 test.beforeEach(async ({ page }) => {
-  await page.goto("http://localhost:3000");
+  await page.goto("/");
 });
 
 test("verify navigation menu is displayed (items, collections, divisions, about", async ({
@@ -66,4 +66,18 @@ test("verify footer links are visible", async ({ page }) => {
   await expect(dchomepage.footerSocialTwitter).toBeVisible();
   await expect(dchomepage.footerSocialInstagram).toBeVisible();
   await expect(dchomepage.footerSocialYouTube).toBeVisible();
+});
+
+test("verify feedback button is visible", async ({ page }) => {
+  const dchomepage = new DCHomepage(page);
+  await expect(dchomepage.feedbackButton).toBeVisible();
+  await dchomepage.feedbackButton.click();
+  await expect(dchomepage.feedbackCommentRadioButton).toBeVisible();
+  await expect(dchomepage.feedbackCorrectionRadioButton).toBeVisible();
+  await expect(dchomepage.feedbackBugRadioButton).toBeVisible();
+  await expect(dchomepage.feedbackTextArea).toBeVisible();
+  await expect(dchomepage.feedbackSubmitButton).toBeVisible();
+  await expect(dchomepage.feedbackCancelButton).toBeVisible();
+  await expect(dchomepage.feedbackPrivacyPolicy).toBeVisible();
+  await dchomepage.feedbackCancelButton.click();
 });