+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Added Image ID to Items page (DR-3695)
+- Added Playwright test feedback button to the home page (DR-3658)
 
 ### Fixed
 - Fixed Order Print button (DR-3486)